Improve filebeat config to drop only unwanted events
Allow filebeat logs to be sent, since the metrics are useful, but limit these to every 5 mins.

Only drop logs from the Rancher ipsec router, and not the other networking components (this is the noisy one). We still have the last 50MB of these on any host.

Drop the Rancher scheduler's 'Updating resource pool' logs, since these are frequent and not an error.